feat: switch to optional dependencies
