+++ package.json
@@ -32,14 +32,16 @@
   },
   "homepage": "https://github.com/wix-incubator/esbuild-jest-cli#readme",
   "dependencies": {
-    "bunyamin": "^1.5.1",
-    "bunyan": "^2.0.0",
-    "bunyan-debug-stream": "^3.1.0",
-    "cosmiconfig": "^8.1.3",
+    "bunyamin": "^1.6.3",
     "find-up": "^7.0.0",
     "import-from": "^4.0.0",
     "lodash.merge": "^4.6.2",
     "resolve-from": "^5.0.0"
+  },
+  "optionalDependencies": {
+    "bunyan": "^2.0.0",
+    "bunyan-debug-stream": "^3.1.0",
+    "cosmiconfig": "^9.0.0"
   },
   "peerDependencies": {
     "esbuild": ">=0.17.4"
@@ -47,6 +49,6 @@
   "devDependencies": {
     "@jest/transform": "^29.6.3",
     "@jest/types": "^29.6.3",
-    "esbuild": "^0.21.4"
+    "esbuild": "^0.24.2"
   }
 }
