+++ utils/logger.mjs
@@ -1,14 +1,9 @@
-import * as fs from 'node:fs';
-import * as path from 'node:path';
 import {
   bunyamin,
   nobunyamin,
   isDebug,
   threadGroups,
-  traceEventStream,
 } from 'bunyamin';
-import { createLogger } from 'bunyan';
-import { create as createDebugStream } from 'bunyan-debug-stream';
 
 const PACKAGE_NAME = 'esbuild-jest-cli';
 
@@ -23,8 +18,6 @@ threadGroups.add({
   maxConcurrency: 100500,
 });
 
-bunyamin.useLogger(createBunyanImpl(), 1);
-
 export const logger = bunyamin.child({ tid: PACKAGE_NAME });
 export const optimizedLogger = isDebug(PACKAGE_NAME)
   ? logger
@@ -35,50 +28,3 @@ const noop = () => {};
 export const optimizeTracing = isDebug(PACKAGE_NAME)
   ? ((f) => f)
   : (() => noop);
-
-function createBunyanImpl() {
-  const logPath = process.env.BUNYAMIN_LOG;
-
-  return createLogger({
-    name: PACKAGE_NAME,
-    streams: [
-      {
-        type: 'raw',
-        level: 'warn',
-        stream: createDebugStream({
-          out: process.stderr,
-          showMetadata: false,
-          showDate: false,
-          showPid: false,
-          showProcess: false,
-          showLoggerName: false,
-          showLevel: false,
-          prefixers: {
-            cat: (value) => String(value).split(',', 1)[0],
-          },
-        }),
-      },
-      ...(logPath
-        ? [
-          {
-            type: 'raw',
-            level: 'trace',
-            stream: traceEventStream({
-              filePath: ensureCleanLog(logPath),
-              threadGroups,
-            }),
-          },
-        ]
-        : []),
-    ],
-  });
-}
-
-function ensureCleanLog(filePath) {
-  fs.mkdirSync(path.dirname(filePath), { recursive: true });
-  if (fs.existsSync(filePath)) {
-    fs.unlinkSync(filePath);
-  }
-
-  return filePath;
-}
